requirements.txt
flask
flask-cors
flask-wtf
flask-talisman
flask-limiter
sqlitecloud
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests
transformers
torch
huggingface_hub
openai
gunicorn

